Creación de roles y permisos

+++ app/Http/Controllers/api/DestinoController.php
@@ -1,36 +1,67 @@
 <?php
 
 namespace App\Http\Controllers\api;
+use App\Models\Destinos;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\DestinoCollection;
-use App\Models\Destinos;
+
 use App\Http\Resources\DestinoResource;  // Importar el recurso DestinoResource
 use App\Http\Requests\StoreDestinoRequest;  // Importar la request StoreDestinoRequest
 use App\Http\Requests\UpdateDestinoRequest; // Importar la request UpdateDestinoRequest
 use Symfony\Component\HttpFoundation\Response; // Importar la clase Response para los códigos de estado HTTP
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests; // Importar el trait AuthorizesRequests para la autorización de políticas
+
 class DestinoController extends Controller
 {
-    //// Muestra todas los destinos
+    use AuthorizesRequests; // Usar el trait AuthorizesRequests para la autorización de políticas
+
+    // Muestra todas las recetas
     public function index(){
-         //return Categoria::all(); // Devuelve todas las destinos
-        return new DestinoCollection(Destinos::all());  // Devuelve todas las destinos como recurso API
+        $this->authorize('Ver destinos');  
+        $destinos = Destinos::with('categoria', 'etiquetas', 'user')->get();
+        return DestinoResource::collection($destinos); // Devuelve todas las recetas como recurso API
     }
 
-    // Muestra un destino a partir de su id
+    // Muestra una receta a partir de su id
     public function show(Destinos $destinos){
-        // return $categoria; // Devuelve la categoria
-        $destinos = $destinos->load('categoria');  // Carga los destinos relacionadas con la categoria
-        return new DestinoResource($destinos);  // Devuelve el destino como recurso API 
+        $this->authorize('Ver destinos');  
+        $destinos = $destinos->load('categoria', 'etiquetas', 'user');
+        return new DestinoResource($destinos); // Devuelve la receta como recurso API 
     }
-     // Actualiza una receta existente
-    public function update(UpdateDestinoRequest $request, Destinos $destino){  // Usar la request UpdateDestinoRequest para validar los datos
+
+    // Almacena una nueva receta 
+    public function store(StoreDestinoRequest $request){  // Usar la request StoreRecetasRequest para validar los datos
+        $this->authorize('Crear destinos');  
+       
+        $destino = $request->user()->destinos()->create($request->all());  // Crear una nueva receta asociada al usuario autenticado
+        $destino->etiquetas()->attach(json_decode($request->etiquetas));  // Asociar las etiquetas a la receta (decodificar el JSON recibido)
+
+        $destino->imagen = $request->file('imagen')->store('recetas','public'); // Almacenar la imagen en el disco 'public' dentro de la carpeta 'recetas'
+        $destino->save(); // Guardar la receta con la ruta de la imagen
+ 
+        // Devolver la receta creada como recurso API con código de estado 201 (creado) 
+        return response()->json(new DestinoResource($destino), Response::HTTP_CREATED); 
+    }
+
+    // Actualiza una receta existente
+    public function update(UpdateDestinoRequest $request, Destinos $destino){  // Usar la request UpdateRecetasRequest para validar los datos
+        $this->authorize('Editar destinos');
+
+        $this->authorize('update', $destino);  // Autorizar la acción usando la política RecetaPolicy
+        
         $destino->update($request->all());  // Actualizar la receta con los datos validados
 
         if($etiquetas = json_decode($request->etiquetas)){  // Si se reciben etiquetas, decodificar el JSON
             $destino->etiquetas()->sync($etiquetas);  // Sincronizar las etiquetas (eliminar las que no están y agregar las nuevas)
         }
+
+        if($request->file('imagen')){  // Si se recibe una imagen
+            $destino->imagen = $request->file('imagen')->store('destinos','public');  // Almacenar la imagen en el disco 'public' dentro de la carpeta 'recetas'
+            $destino->save(); // Guardar la receta con la ruta de la imagen
+        }
+        
 
         // Devolver la receta actualizada como recurso API con código de estado 200 (OK)
         return response()->json(new DestinoResource($destino), Response::HTTP_OK);
@@ -38,6 +69,10 @@ class DestinoController extends Controller
 
     // Elimina una receta existente
     public function destroy(Destinos $destino){  // Inyectar la receta a eliminar
+        $this->authorize('Eliminar destinos');
+        
+        $this->authorize('delete', $destino);  // Autorizar la acción usando la política RecetaPolicy
+        
         $destino->delete();  // Eliminar la receta
 
         // Devolver una respuesta vacía con código de estado 204 (No Content)
